The bug identified
take a look at the example. you would realize that the task 1 and 2 are supposed to be swapping from one to the other. but if you run it, you will notice it gets struck at a point. initially Author thought this was a bug, but apparently it was a synchronization issue that was supposed to happen at one point or the other. the issue was : sometimes just as task1 calls release(2), before task1 runs `task_exit()` it runs release(2). hence after task2 finishes and calls task1 task1 continues from where it left, i.e the task_exit method is executed. which ends the task, hence the whole loop of task1 and task2 ends. This would affect event management and semaphores, should be considered a critical issue.